% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{diagnostic_plots}
\alias{diagnostic_plots}
\title{Generate Diagnostic Plots for a Logistic Regression Model}
\usage{
diagnostic_plots(model)
}
\arguments{
\item{model}{An object of class \code{"simple_logistic"}.}
}
\value{
A list containing ggplot objects for the diagnostic plots:
\item{residuals_vs_fitted}{Plot of residuals versus fitted values.}
\item{roc_curve}{Receiver Operating Characteristic (ROC) curve.}
}
\description{
Creates diagnostic plots, including residuals vs fitted values and ROC curve.
}
\details{
This function helps assess the fit of the logistic regression model by providing
common diagnostic plots.
}
\examples{
library(SimpleLogistic)
data("health_data")
model <- simple_logistic(outcome ~ age + bmi + treatment, data = health_data)
plots <- diagnostic_plots(model)
print(plots$residuals_vs_fitted)
print(plots$roc_curve)

}
