% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prediction.R
\name{plot_predictions}
\alias{plot_predictions}
\title{Plot Predicted Probabilities Against Actual Outcomes}
\usage{
plot_predictions(model, newdata = NULL)
}
\arguments{
\item{model}{An object of class \code{"simple_logistic"}.}

\item{newdata}{(Optional) A new data frame for which to make predictions.}
}
\value{
A ggplot object showing the distribution of predicted probabilities.
}
\description{
This function creates a histogram of predicted probabilities, colored by actual outcomes.
}
\details{
This plot helps visualize how well the model's predicted probabilities align with the actual outcomes.
}
\examples{
library(SimpleLogistic)
data("health_data")
model <- simple_logistic(outcome ~ age + bmi + treatment, data = health_data)
plot_predictions(model)

}
